HW21

+++ Src/01-Domain/Services/App.Domain.Services.AppService/CarModelAppSer/CarModelAppService.cs
@@ -1,6 +1,7 @@
 ﻿using App.Domain.Core.Cars.AppService;
 using App.Domain.Core.Cars.Entity;
 using App.Domain.Core.Cars.Service;
+using App.Domain.Core.SerResult;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -27,20 +28,37 @@ namespace App.Domain.Services.AppService.CarModelAppSer
         {
             return _carModelService.GetModelById(id);
         }
-
-        public void CreateModel(CarModel carModel)
+        public ServiceResult CreateModel(CarModel carModel)
         {
+            if (string.IsNullOrWhiteSpace(carModel.Name))
+            {
+                return new ServiceResult(false, "نام مدل خودرو نمی‌تواند خالی باشد.");
+            }
+
             _carModelService.CreateModel(carModel);
+            return new ServiceResult(true, "مدل خودرو با موفقیت اضافه شد.");
         }
 
-        public void UpdateModel(CarModel carModel)
+        public ServiceResult UpdateModel(CarModel carModel)
         {
+            if (_carModelService.IsCarModelUsed(carModel.Id))
+            {
+                return new ServiceResult(false, "این مدل خودرو در سیستم استفاده شده است و نمی‌توان آن را ویرایش کرد.");
+            }
+
             _carModelService.UpdateModel(carModel);
+            return new ServiceResult(true, "مدل خودرو ویرایش شد.");
         }
 
-        public void DeleteModel(int id)
+        public ServiceResult DeleteModel(int id)
         {
+            if (_carModelService.IsCarModelUsed(id))
+            {
+                return new ServiceResult(false, "این مدل خودرو در سیستم استفاده شده است و نمی‌توان آن را حذف کرد.");
+            }
+
             _carModelService.DeleteModel(id);
+            return new ServiceResult(true, "مدل خودرو حذف شد.");
         }
     }
 }
